chore(cleanup): consolidate and cleanup scripts
