+++ .gitignore
@@ -1,3 +1,4 @@
 # data folders
 **/csv/
 **/results/
+**/rawdata/
