+++ README.md
@@ -2,15 +2,19 @@
 
 ## What is it?
 
-This repository contains a couple of Python scripts that automate the creation of daily fantasy football rosters.
+This repository contains a couple of Python scripts that automate the creation of daily fantasy football rosters. Designed to maximize the provided raw data, this application was created with 1v1 and small contests in mind.
 
-Currently supports Yahoo main slate matches and OwnersBox Thursday -> Monday games.
+Currently supports Yahoo main slate Sunday matches and OwnersBox Thursday thru Monday games.
 
 ## How to use:
 
-`> python fetch_data.py`
+`> python fetch_data.py <format>`
 
-`> python dfs-solver.py <file>`
+`> python dfs-solver.py <file> <format>`
+
+Where format is either `--ownersbox` || `-o`, or `--yahoo` || `-y`. If not specified it will default to Yahoo format.
+
+Where file is the path to the rawdata `.csv` file. 
 
 ## Example output:
 ```
@@ -33,7 +37,7 @@ Projected points: 130.9
 
 ![2021](https://user-images.githubusercontent.com/10425301/195481619-48b12822-7c6a-4bd4-a84a-4b0e32d80f1d.png)
 
-Yahoo Diamond Rank (top 1%)
+Yahoo Diamond Rank (99th Percentile)
 
 Placed 504/223186 in the 2021 Yahoo cup.
 
@@ -41,6 +45,6 @@ Placed 504/223186 in the 2021 Yahoo cup.
 
 ![2020](https://user-images.githubusercontent.com/10425301/124058137-c974f680-d9f6-11eb-9a15-3876a86e101e.png)
 
-Yahoo Diamond Rank (top 5%)
+Yahoo Diamond Rank (95th Percentile)
 
 Placed 972/142,492 in the BetMGM 2020 Yahoo Cup
